backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic>=2.10.0,<3.0.0
pydantic-ai[mcp]==0.0.21
python-dotenv==1.0.0
requests>=2.31.0
python-multipart==0.0.6
httpx>=0.24.0 